+++ CHANGES.md
@@ -11,6 +11,7 @@ Running 4.3.1
 * [Account][Security] Handle comparison between a field and 'null' value
 * [Vega] Add healthCheck readiness (no subcomponents RED)
 * [Datastore][S3] InputStream in S3FileStorePlugin no longer shared (may lead to corrupt file on subsequent reads)
+* [Datastore][S3] Add health check
 * [Vega][Json] Don't set null values to object properties if not present in the json (fail if the attribute is a list)